requirements.txt
appdirs==1.4.4
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.8b0
cattrs==1.8.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
decorator==4.4.2
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
lxml==4.6.3
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.3.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-metallum==1.1.0
PyYAML==5.4.1
regex==2021.8.28
requests==2.25.1
requests-cache==0.7.4
six==1.16.0
soupsieve==2.2
spotipy==2.19.0
tomli==1.2.1
traitlets==5.0.5
typer==0.4.0
typing-extensions==3.10.0.2
url-normalize==1.4.3
urllib3==1.26.3
wcwidth==0.2.5
